requirements.txt
mail-parser
requests
python-dotenv